updated table models as per the feedback and clean the code

+++ src/models/user.model.js
@@ -7,16 +7,16 @@ export default class User extends BaseModel {
     type: 'object',
 
     properties: {
-      id: { type: 'integer' },
+      id: { type: 'bigInteger' },
       userName: { type: 'string', maxLength: 255 },
       createdAt: { type: 'string', format: 'date-time' },
       updatedAt: { type: 'string', format: 'date-time' },
-      firstName: { type: 'string', minLength: 1, maxLength: 255 },  
-      lastName:{ type: 'string', minLength: 1, maxLength: 255 },
+      firstName: { type: 'string', minLength: 1, maxLength: 255 },
+      lastName: { type: 'string', minLength: 1, maxLength: 255 },
       email: { type: 'string', minLength: 1, maxLength: 255, format: 'email' },
-      workspaceId: { type: 'string'}
+      workspaceId: { type: 'string' }
     },
 
-    required: ['firstName', 'lastName','email'],
+    required: ['firstName', 'lastName', 'email'],
   };
 }
